# perf: Add full program benchmark for kustomize build