@shapirus 2023-10-30
One important aspect that I'd like to point out here is that the benchmarks should cover the following cases:

- single invocation of kustomize to process a big and complex resource tree
- multiple invocations of kustomize to process a small set of manifests on each run (simulates e.g. argocd handling many microservices, or a validation CI job running on the repository hosting a large set of manifests)

These cases are very different in terms of the performance impact that specific parts of code create, yet each of them is quite valid as far as practical usage is concerned.